+++ src/main/java/lotto/step2/domain/Money.java
@@ -1,4 +1,4 @@
-package lotto.step2;
+package lotto.step2.domain;
 
 import java.math.BigDecimal;
