+++ src/test/java/lotto/step2/RankTest.java
@@ -1,5 +1,6 @@
 package lotto.step2;
 
+import lotto.step2.domain.Rank;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
